Allow high values of dlType in Rule condition.
This addresses #537. The dlType in Rule/Condition was previously
a Short, so values above 32K were not allowed (for example 0x86DD
for IPv6).

Added changes to unit tests to cover this change.

+++ midonet-client/src/main/java/org/midonet/client/resource/Rule.java
@@ -193,7 +193,7 @@ public class Rule extends ResourceBase<Rule, DtoRule> {
      *
      * @return
      */
-    public Short getDlType() {
+    public Integer getDlType() {
         return principalDto.getDlType();
     }
 
@@ -590,8 +590,8 @@ public class Rule extends ResourceBase<Rule, DtoRule> {
      * @param dlType
      * @return this
      */
-    public Rule dlType(Short dlType) {
-        principalDto.setDlType(dlType);
+    public Rule dlType(Integer dlType) {
+        principalDto.setDlType(dlType.shortValue());
         return this;
     }
